requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0
psycopg2>=2.7.5
